% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AGQfuns.R
\name{fit_gen}
\alias{fit_gen}
\title{fit GLMM for varying values}
\usage{
fit_gen(data, formula, family, cluster = NULL, pkg = c("lme4", "glmmML",
  "glmmPQL"), maxAGQ = 10, AGQvec = sseq(maxAGQ), verbose = TRUE,
  truevals = NULL)
}
\arguments{
\item{data}{}

\item{formula}{model formula (fixed-effect or full, depending on pkg)}

\item{cluster}{cluster variable specification (for glmmPQL or glmmML)}

\item{pkg}{package to use for fitting}

\item{maxAGQ}{maximum AGQ value (0=PQL)}

\item{AGQvec}{vector of specified AGQ orders}

\item{verbose}{print verbose output}

\item{truevals}{vector of true parameter values}
}
\description{
fit GLMM for varying values
}

