% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AGQfuns.R
\name{rmsetab_to_combdf}
\alias{rmsetab_to_combdf}
\title{compute RMSE and return a reduced table}
\usage{
rmsetab_to_combdf(S, v, std = NULL, ret = "rmse", dims = 2:3,
  calc.range = TRUE)
}
\arguments{
\item{S}{a 4-dimensional numeric array: { sim * AGQ * vars * type }}
}
\description{
compute RMSE and return a reduced table
}

