% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AGQfuns.R
\name{sim_to_rmsetab}
\alias{sim_to_rmsetab}
\title{compute FE and RE summaries}
\usage{
sim_to_rmsetab(x, v, std = NULL, ret = "rmse")
}
\arguments{
\item{x}{a 3-dimensional numeric array: { AGQ * vars * type }}

\item{v}{name of RE variance}

\item{std}{how to standardize}

\item{ret}{what to return}
}
\description{
compute FE and RE summaries
}

