% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AGQfuns.R
\name{sumfun2}
\alias{sumfun2}
\title{computes summary statistics for}
\usage{
sumfun2(fitted, truevals = NULL)
}
\arguments{
\item{fitted}{a fitted GLMM model (glmer, glmmML, or glmmPQL)}

\item{truevals}{optional vector of true parameter values (beta, RE sd)}
}
\description{
computes summary statistics for
}

