% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AGQfuns.R
\name{vcov.VarCorr.merMod}
\alias{vcov.VarCorr.merMod}
\title{Wald variance-covariance matrix of lme4 RE parameters}
\usage{
\method{vcov}{VarCorr.merMod}(fit, ...)
}
\arguments{
\item{fit}{an lme4 fit}
}
\description{
Wald variance-covariance matrix of lme4 RE parameters
}
\details{
by brute force, finds the Wald var-cov matrix
of lme4 RE parameters (on the 'user', i.e. sd-corr-sigma scale)
}

